Move Web configs to run_config dir
