+++ Web/src/components/Body/Feed/Post/Post.jsx
@@ -1,7 +1,7 @@
 // eslint-disable-next-line no-unused-vars
 import React, {Component} from 'react'
 import './PostStyles.scss'
-import API from '@/apiConfig'
+import API from '@/run_config/apiConfig'
 
 class Post extends Component {
   componentWillMount () {
